video2text/pyproject.toml
[project]
name = "video2text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20240930",
]
